refactor: renomear, mover e atualizar as rotas dos arquivos
