+++ meu-app/frontend/app/screens/recoverYourAccount.jsx
@@ -5,7 +5,7 @@ import Button from "../components/Button";
 import useRecConta from "../hooks/useRecConta";
 import EmailInput from "../components/EmailInput";
 // arquivo config da API
-import api from "../../services/api";
+import api from "../services/api";
 import AsyncStorage from "@react-native-async-storage/async-storage";
 
 function RecConta() {
